+++ packages/common/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @u-moni/common
 
+## 1.6.0
+
+### Minor Changes
+
+- support React project
+
 ## 1.5.0
 
 ### Minor Changes