+++ packages/recordScreen/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @u-moni/recordscreen
 
+## 2.2.0
+
+### Minor Changes
+
+- support React project
+
+### Patch Changes
+
+- Updated dependencies
+  - @u-moni/common@1.6.0
+
 ## 2.1.10
 
 ### Patch Changes